refactore geomodel_api
