+++ src/bin/utilities/ringmesh-stats.cpp
@@ -41,7 +41,7 @@
 #include <ringmesh/basic/command_line.h>
 #include <ringmesh/geomodel/core/entity_type_manager.h>
 #include <ringmesh/geomodel/core/geomodel.h>
-#include <ringmesh/geomodel/tools/geomodel_api.h>
+#include <ringmesh/geomodel/core/geomodel_api.h>
 #include <ringmesh/io/io.h>
 
 /*!
