+++ src/bin/utilities/ringmesh-tetrahedralize.cpp
@@ -40,7 +40,7 @@
 
 #include <ringmesh/basic/command_line.h>
 #include <ringmesh/geomodel/core/geomodel.h>
-#include <ringmesh/geomodel/tools/geomodel_api.h>
+#include <ringmesh/geomodel/tools/geomodel_tools.h>
 #include <ringmesh/io/io.h>
 
 /*!
