+++ src/ringmesh/geomodel/builder/stratigraphic_column_builder.cpp
@@ -35,7 +35,7 @@
 
 #include <ringmesh/geomodel/builder/stratigraphic_column_builder.h>
 #include <ringmesh/geomodel/core/geomodel_geological_entity.h>
-#include <ringmesh/geomodel/tools/geomodel_api.h>
+#include <ringmesh/geomodel/core/geomodel_api.h>
 #include <tinyxml2.h>
 
 namespace RINGMesh
