+++ src/ringmesh/geomodel/core/CMakeLists.txt
@@ -35,6 +35,7 @@ target_sources(${target_name}
     PRIVATE
         "${lib_source_dir}/common.cpp"
         "${lib_source_dir}/entity_type_manager.cpp"
+        "${lib_source_dir}/geomodel_api.cpp"
         "${lib_source_dir}/geomodel_entity.cpp"
         "${lib_source_dir}/geomodel_geological_entity.cpp"
         "${lib_source_dir}/geomodel_mesh_entity.cpp"
@@ -46,6 +47,7 @@ target_sources(${target_name}
         "${lib_include_dir}/common.h"
         "${lib_include_dir}/entity_type_manager.h"
         "${lib_include_dir}/entity_type.h"
+        "${lib_include_dir}/geomodel_api.h"
         "${lib_include_dir}/geomodel_entity.h"
         "${lib_include_dir}/geomodel_geological_entity.h"
         "${lib_include_dir}/geomodel_mesh_entity.h"
