+++ src/ringmesh/geomodel/tools/CMakeLists.txt
@@ -34,13 +34,13 @@
 target_sources(${target_name}
     PRIVATE
         "${lib_source_dir}/common.cpp"
-        "${lib_source_dir}/geomodel_api.cpp"
+        "${lib_source_dir}/geomodel_tools.cpp"
         "${lib_source_dir}/geomodel_repair.cpp"
         "${lib_source_dir}/geomodel_validity.cpp"
         "${lib_source_dir}/mesh_quality.cpp"
     PRIVATE # Could be PUBLIC from CMake 3.3
         "${lib_include_dir}/common.h"
-        "${lib_include_dir}/geomodel_api.h"
+        "${lib_include_dir}/geomodel_tools.h"
         "${lib_include_dir}/geomodel_repair.h"
         "${lib_include_dir}/geomodel_validity.h"
         "${lib_include_dir}/mesh_quality.h"
