+++ src/ringmesh/io/geomodel_builder_gocad.cpp
@@ -39,7 +39,7 @@
 #include <ringmesh/io/geomodel_builder_gocad.h>
 #include <ringmesh/geomodel/core/geomodel.h>
 #include <ringmesh/geomodel/core/geomodel_mesh_entity.h>
-#include <ringmesh/geomodel/tools/geomodel_api.h>
+#include <ringmesh/geomodel/core/geomodel_api.h>
 #include <ringmesh/mesh/mesh.h>
 #include <ringmesh/mesh/mesh_index.h>
