+++ src/ringmesh/io/io_geomodel.cpp
@@ -50,7 +50,7 @@
 #include <ringmesh/io/geomodel_builder_gocad.h>
 #include <ringmesh/geomodel/core/geomodel.h>
 #include <ringmesh/geomodel/core/geomodel_mesh_entity.h>
-#include <ringmesh/geomodel/tools/geomodel_api.h>
+#include <ringmesh/geomodel/core/geomodel_api.h>
 #include <ringmesh/geomodel/tools/geomodel_validity.h>
 #include <ringmesh/geomodel/core/well.h>
